Adds necessary functionality for iOS test running
The dependency on Moq has been (temporarily?) removed.  It seems to be
incompatible with the iOS simulator--probably the device, too.

FramePixelData has been introduced as a way to abstract image data
access, image saving and image loading.  MonoTouch does not implement
System.Drawing.Imaging, so a platform-specific solution was needed.

ConsoleInterface has been broken up significantly to better share
functionality with MobileInterface.

Upgraded to NUnit 2.6, which eliminates a dependency on
System.Configuration, which MonoTouch does not implement.