Drastic measures to simplify testing
Creating components for tests is no longer the preffered way to create
simple visual tests.  Instead, prefer Game.DrawWith, and the other *With
events exposed by TestGameBase.  Follow that with a call to
RunSingleFrameTest or RunMultiFrameTest.

Captured image folders and filenames are now automatically generated in
all cases.

Reduced the size of some reference images (that picture of outer space,
even when PNG compressed, was consuming quite a bit of space).

Reorganized reference image names to be more natural:
Clear/frame-01.png => Clear-1.png, for example.

Improved tests.xsl with better handling of parameterized tests and
generally better layout and display.

Adds a few tests for GraphicsDevice.Viewport for issue #329