Adds a framework for interactive tests (iOS only)
To add a new interactive test, subclass Microsoft.XNA.Framework.Game and
decorate the subclass with [InteractiveTest].  Doing this will add your
interactive test to the test launching GUI.  Be sure to provide a way to
exit the game (a button, say) so users can continue running other tests
when they've finished one.

An extremely crude GUI framework for writing interactive tests has also
been added.  It is somewhat verbose, at the moment, but still easier
than doing it all by hand.

Adds a test for the Guide class, and a crude test for the crude test GUI
framework.